+++ foxes/models/turbine_types/__init__.py
@@ -1,3 +1,4 @@
 from .PCt_file import PCtFile
 from .PCt_two_files import PCtTwoFiles
 from .null_type import NullType
+from .wsrho2PCt_two_files import WsRho2PCtTwoFiles